# Store NonbondedPairList scales as RealType