--- a/timemachine/cpp/src/kernels/k_chiral_restraint.cuh
+++ b/timemachine/cpp/src/kernels/k_chiral_restraint.cuh
@@ -9,7 +9,7 @@ void __global__ k_chiral_atom_restraint(
     const int R, // number of restraints
     const double *__restrict__ coords,
     const double *__restrict__ params, // [R]
-    const int *__restrict__ idxs,      // [R, 2]
+    const int *__restrict__ idxs,      // [R, 4]
     unsigned long long *__restrict__ du_dx,
     unsigned long long *__restrict__ du_dp,
     __int128 *__restrict__ u) {