--- a/timemachine/cpp/src/kernels/k_nonbonded_pair_list.cuh
+++ b/timemachine/cpp/src/kernels/k_nonbonded_pair_list.cuh
@@ -22,7 +22,7 @@ void __global__ k_nonbonded_pair_list(
     const double *__restrict__ params,
     const double *__restrict__ box,
     const int *__restrict__ pair_idxs,   // [M, 2] pair-list of atoms
-    const RealType *__restrict__ scales, // [M]
+    const RealType *__restrict__ scales, // [M, 2]
     const double beta,
     const double cutoff,
     unsigned long long *__restrict__ du_dx,