* Had this lying around for awhile, can't remember if there is some issue with it.
* Minor doc update that I noticed as part of reviewing https://github.com/proteneer/timemachine/pull/1243