@proteneer 2024-02-03
I think by default we prefer doubles everywhere unless there's a speed benefit to converting to float. Does it make vacuum simulations faster?