@badisa 2024-02-04
> I think by default we prefer doubles everywhere unless there's a speed benefit to converting to RealType. 
In the case that we only read the values (and convert from double to float https://github.com/proteneer/timemachine/blob/master/timemachine/cpp/src/kernels/k_nonbonded_pair_list.cuh#L99-L100) and don't write them, is that the case? We are storing 2x the number of bytes and doing a conversion that doesn't impact behavior.

> Does it make vacuum simulations faster?

Good question, realized that we don't have a vacuum simulation in our benchmarks. Will add