@badisa 2024-02-04
@proteneer Added a vacuum benchmark in https://github.com/proteneer/timemachine/pull/1245/commits/3ba48f8308c08f1824de63740b660affd08f130b

Seems to be about 2% faster with the change, but running this on my personal laptop so may not be converged

## Before
vacuum-rbfe: N=35 speed: 12610.71ns/day dt: 2.5fs (ran 10000000 steps in 171.32s)

## After
vacuum-rbfe: N=35 speed: 12837.65ns/day dt: 2.5fs (ran 10000000 steps in 168.29s)