@proteneer 2024-02-04
Wow! 12.8ms per day - I wonder why our vacuum simulations are so slow. Is this with or without HREX?