@badisa 2024-02-04
> Wow! 12.8us per day - I wonder why our vacuum simulations are so slow. Is this with or without HREX? (Guessing this is single state simulations)

Without HREX. Can also look at the HREX tests, though more going on in HREX so didn't seem as clean of a benchmark.