@badisa 2024-02-05
Seems to make no difference on Tesla cards which is target, closing. Will move over the doc changes to another PR at some point (perhaps deboggling)